Comments

+++ Labb3b/src/UserInterface.java
@@ -1,3 +1,8 @@
+/** Objects of this class represents a UserInteraface 
+ *	So that  a user can interact with the object CollectionOfBooks
+ */
+
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -13,25 +18,33 @@ public class UserInterface {
 		
 	}
 
+	
+	/** This method is the main running loop that you call when you want to run the user interface
+	 */
 	public void run(){
-		try{
-			loadFile();
-		}
-		catch(ClassNotFoundException e){
-			System.out.println("Currupted savefile all data is lost");
-		}
-		catch(FileNotFoundException e){
-			System.out.println("No save file found!");
-		}
-		catch(IOException e){
-			System.out.println("Error while loading file, all data is lost");
+		System.out.print("Do you want to load a file?(1/0)");
+		if(scan.nextInt() == 1){
+			try{
+				loadFile();
+			}
+			catch(ClassNotFoundException e){
+				System.out.println("Currupted savefile all data is lost");
+			}
+			catch(FileNotFoundException e){
+				System.out.println("Save file not found, creating new.");
+			}
+			catch(IOException e){
+				System.out.println("Error while loading file, all data is lost");
+			}
 		}
 		while(running){
 			menu();
 		}
-		
+
 	}
-	
+
+	/*The main menu of the interface where all choises are made
+	*/
 	private void menu(){
 		int choice;
 		
@@ -40,7 +53,8 @@ public class UserInterface {
 		System.out.println("2. Remove Book");
 		System.out.println("3. Search Books");
 		System.out.println("4. List all Books");
-		System.out.println("5. Save and Exit");
+		System.out.println("5. Sort books");
+		System.out.println("6. Save and Exit");
 		choice = scan.nextInt();
 		switch(choice){
 		case 1:
@@ -56,6 +70,9 @@ public class UserInterface {
 			listBooks();
 			break;
 		case 5:
+			database.sort(database.getList());
+			break;
+		case 6:
 			saveAndExit();
 			break;
 		default:
@@ -65,6 +82,8 @@ public class UserInterface {
 
 	}
 	
+	/*The Search menu where you enter your search word and what to search after
+	*/
 	private void searchMenu(){
 		int choice;
 		
@@ -74,17 +93,17 @@ public class UserInterface {
 		System.out.println("3. Search ISBN");
 		System.out.println("4. Exit");
 		choice = scan.nextInt();
-		
+		scan.nextLine();
 		System.out.print("Enter Search Word: ");
 		switch(choice){
 		case 1:
-			getBooksByTitle(scan.next());
+			getBooksByTitle(scan.nextLine());
 			break;
 		case 2:
-			getBooksByAuthor(scan.next());
+			getBooksByAuthor(scan.nextLine());
 			break;
 		case 3:
-			getBooksByIsbn(scan.next());
+			getBooksByIsbn(scan.nextLine());
 			break;
 		case 4:
 			break;
@@ -95,6 +114,8 @@ public class UserInterface {
 
 	}
 	
+	/*Adds a b00k in the Collectionofbooks object
+	*/
 	private void addBook(){
 		String titel;
 		String isbn;
@@ -102,45 +123,57 @@ public class UserInterface {
 		double price;
 		String author;		
 		
+		scan.nextLine();
 		System.out.println("---Add Book---");
 		System.out.println("Enter Titel: ");
-		titel = scan.next();
+		titel = scan.nextLine();
 		System.out.println("Enter ISBN: ");
-		isbn = scan.next();
+		isbn = scan.nextLine();
 		System.out.println("Enter Edition: ");
 		edition = scan.nextInt();
 		System.out.println("Enter Price: ");
 		price = scan.nextDouble();
+		scan.nextLine();
 		System.out.println("Enter Author: ");
-		author = scan.next();
+		author = scan.nextLine();
 		
 		database.addBook(new Book(titel, isbn, edition, price, author));
 		System.out.println("Add Additional Athours? (1/0)");
 		while(scan.nextInt() == 1){
+			scan.nextLine();
 			System.out.println("Enter Author: ");
-			database.getBook(database.size()-1).addAuthor(scan.next());
+			database.getBook(database.size()-1).addAuthor(scan.nextLine());
 			
 			System.out.println("Add Additional Athours? ");
 		}
 		System.out.println("Book Added");
 	}
 	
+	/*Asks for the id and removes the book
+	*/
 	private void removeBook(){
 		listBooks();
 		int choice;
 		System.out.println("Which book do you want to remove: ");
 		choice = scan.nextInt();
-		System.out.println(database.getBook(choice).toString());
-		System.out.println("Delete: (1/0) ");
-		if(scan.nextInt() == 1){
-			database.removeBook(choice);
-			System.out.println("Book Removed");
+		if(choice < database.size()){
+			System.out.println(database.getBook(choice).toString());
+			System.out.println("Delete: (1/0) ");
+			if(scan.nextInt() == 1){
+				database.removeBook(choice);
+				System.out.println("Book Removed");
+			}
+			else{
+				System.out.println("Book not removed");
+			}	
 		}
 		else{
-			System.out.println("Book not removed");
-		}	
+			System.out.println("Book does not exist");
+		}
 	}
 	
+	/*Lists all of the books inside of the collection
+	*/
 	private void listBooks(){
 		System.out.println("---Books in Database---\n");
 		
@@ -150,6 +183,8 @@ public class UserInterface {
 		
 	}
 	
+	/*Prints the search results
+	*/
 	private void printSearch(ArrayList<Book> searchResults){
 		System.out.println("---Search Results---");
 
@@ -158,6 +193,8 @@ public class UserInterface {
 		}
 	}
 	
+	/*Sends the search word to the right method depending on what you want to search after
+	*/
 	private void getBooksByTitle(String title){
 		printSearch(database.searchTitel(title));		
 	}
@@ -170,9 +207,15 @@ public class UserInterface {
 		printSearch(database.searchAuthor(author));
 	}
 			
+	
+	/*Saves the database object to a file
+	*/
 	private void saveAndExit(){
+		String Name;
+		System.out.print("Name of save: ");
+		Name = scan.next();
 		try{
-		database.saveToFile();
+		database.saveToFile(Name);
 		}
 		catch(IOException e){
 			System.out.println("Error while saving file\n try again? (1/0)");
@@ -183,8 +226,13 @@ public class UserInterface {
 		running = false;		
 	}
 	
+	
+	/*Loads a object of CollectionOfBooks and stores it in database.
+	*/
 	private void loadFile() throws IOException, ClassNotFoundException{
-		database.loadFromFile();
+		System.out.print("Name of save: ");
+
+		database.loadFromFile(scan.next());
 	}
